+++ CHANGELOG.markdown
@@ -2,6 +2,7 @@
 
 ## Unreleased
 
+* Raise and log on unknown server paths. (Tristan Dunn)
 * Update dependencies. (Tristan Dunn)
 
 ## 0.6.0 — January 23, 2013